# Add sscal.c + microkernels for Haswell, Zen, Skylake and newer.