--- a/kernel/x86_64/sscal_microk_haswell-2.c
+++ b/kernel/x86_64/sscal_microk_haswell-2.c
@@ -111,7 +111,7 @@ static void sscal_kernel_16( BLASLONG n, FLOAT *alpha, FLOAT *x)
 	  "%xmm0", "%xmm1", "%xmm2", "%xmm3",
 	  "%xmm4", "%xmm5", "%xmm6", "%xmm7",
 	  "%xmm8", "%xmm9", "%xmm10", "%xmm11",
-	  "%xmm12", "%xmm13", "%ymm14", "%xmm15",
+	  "%xmm12", "%xmm13", "%xmm14", "%xmm15",
 	  "memory"
 	);