@bartoldeman 2022-12-06
I saw the best performance increase on Sky Lake (Xeon Gold 6148 CPU):
old:
```
$ OPENBLAS_NUM_THREADS=1 OPENBLAS_LOOPS=1000000 ./sscal.goto.slow 2048 16384 2048

From : 2048  To : 16384 Step = 2048 Inc_x = 1 Inc_y = 1 Loops = 1000000
   SIZE       Flops
   2048 :     9443.98 MFlops   0.000000 sec
   4096 :     9378.81 MFlops   0.000000 sec
   6144 :     9434.39 MFlops   0.000001 sec
   8192 :     9238.11 MFlops   0.000001 sec
  10240 :     6727.11 MFlops   0.000002 sec
  12288 :     6721.38 MFlops   0.000002 sec
  14336 :     6731.90 MFlops   0.000002 sec
  16384 :     6725.62 MFlops   0.000002 sec

```
new
```
From : 2048  To : 16384 Step = 2048 Inc_x = 1 Inc_y = 1 Loops = 1000000
   SIZE       Flops
   2048 :    30899.86 MFlops   0.000000 sec
   4096 :    31525.22 MFlops   0.000000 sec
   6144 :    33467.99 MFlops   0.000000 sec
   8192 :    31956.93 MFlops   0.000000 sec
  10240 :    11018.20 MFlops   0.000001 sec
  12288 :    11027.15 MFlops   0.000001 sec
  14336 :    10999.59 MFlops   0.000001 sec
  16384 :    11041.60 MFlops   0.000001 sec
```

Zen and Kaby Lake also saw good improvements. Similar to #3848 there's a slight decrease in performance on Broadwell (E5-2683 v4) in the L2 cache.

old
```
From : 2048  To : 16384 Step = 2048 Inc_x = 1 Inc_y = 1 Loops = 1000000
   SIZE       Flops
   2048 :     5539.16 MFlops   0.000000 sec
   4096 :     6847.83 MFlops   0.000001 sec
   6144 :     7334.60 MFlops   0.000001 sec
   8192 :     7356.24 MFlops   0.000001 sec
  10240 :     6351.74 MFlops   0.000002 sec
  12288 :     6355.27 MFlops   0.000002 sec
  14336 :     6362.03 MFlops   0.000002 sec
  16384 :     6361.14 MFlops   0.000003 sec
```
new
```
From : 2048  To : 16384 Step = 2048 Inc_x = 1 Inc_y = 1 Loops = 1000000
   SIZE       Flops
   2048 :    12901.67 MFlops   0.000000 sec
   4096 :    15697.27 MFlops   0.000000 sec
   6144 :    17798.50 MFlops   0.000000 sec
   8192 :    18112.80 MFlops   0.000000 sec
  10240 :     6102.99 MFlops   0.000002 sec
  12288 :     6110.27 MFlops   0.000002 sec
  14336 :     6111.52 MFlops   0.000002 sec
  16384 :     6111.83 MFlops   0.000003 sec
```