@martin-frbg 2022-12-06
Surprisingly effective. We got one CI failure from (probably) using an ancient version of gcc, but I guess this could be worked around by conditionally falling back to the old kernel in the respective KERNEL files. (Or abandoning gcc 4.8, last released 7 years ago)